docs: Add tutorial README with learning path structure (#177)
Add tutorial index/README that provides a roadmap for the NeMo Gym
tutorial series.

This establishes the structure for upcoming tutorial documentation,
organized into:
- Getting Started
- Training
- Resource Servers
- Advanced Operations

- This PR includes links to tutorials that will be added in follow-up
PRs. The structure is intentionally created first to provide a clear
roadmap for reviewers and users.
- Did not include previously considered sections for models and agents
since we aren't planning those tutorials for 0.1.0 release.

Related to #119

Signed-off-by: Chris Wing <cwing@nvidia.com>